Add a build step that merges the generated Java API into a single jar
This jar is excutable on MacOS and Linux. As the swig tool generates different api files for Linux and MacOS,
but the build tool included only the Linux java files, the cvc4-turnkey jar failed for certain instructions on MacOS.

The new build step merges the Rationa.java, Integer.java and CVC4JNI.java in a way that allows execution on MacOS and Linux.